+++ src/imio/smartweb/core/browser/sitemap.py
@@ -4,7 +4,6 @@ from imio.smartweb.core.config import EVENTS_URL
 from imio.smartweb.core.config import NEWS_URL
 from imio.smartweb.core.contents.rest.utils import get_auth_sources_response
 from imio.smartweb.core.contents.rest.utils import get_entity_id
-from imio.smartweb.core.utils import get_json
 from plone import api
 from plone.base.interfaces import IPloneSiteRoot
 from plone.app.layout.sitemap.sitemap import SiteMapView
@@ -12,6 +11,10 @@ from plone.registry.interfaces import IRegistry
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import normalizeString
 from zope.component import getUtility
+
+import logging
+
+logger = logging.getLogger("imio.smartweb.core")
 
 
 class CustomSiteMapView(SiteMapView):
@@ -122,8 +125,12 @@ class CustomSiteMapView(SiteMapView):
                 return self.request.response.setStatus(
                     404, "No default authentic source found"
                 )
-            brains = api.content.find(UID=auth_source_uid)
-            obj = brains[0].getObject()
+            obj = api.content.get(UID=auth_source_uid)
+            if obj is None:
+                logger.warning(
+                    f"Seems that a main authentic view (for {auth_source_key}) is not found"
+                )
+                continue
             auth_source_view_url = obj.absolute_url()
             results = get_auth_sources_response(
                 auth_source_key, normalizeString(entity_id), (60 * 60 * 24)