+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.2.42 (unreleased)
+1.2.43 (unreleased)
 -------------------
 
 - WEB-4098 : Add affiche scale for section on table view display and one element by lot 
@@ -10,6 +10,16 @@ Changelog
 
 - WEB-4098 : Add new profile for imio.smartweb.orientation to section files
   [thomlamb]
+
+- SUP-35100 : Fix sitemap.xml.gz generation. When a "main" rest view was removed, continue to build sitemap for others rest views
+  [boulch]
+
+
+1.2.42 (2024-04-12)
+-------------------
+
+- SUP-36564 : Fix arcgis external content plugin
+  [boulch]
 
 
 1.2.41 (2024-04-04)